functions/timer_trigger_forecast_api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
ecmwf-opendata
azure-identity>=1.14.0
azure-storage-file-datalake>=12.9.1